requirements.txt
apache-airflow-providers-postgres
pandas
scikit-learn
xgboost
psycopg2-binary
sqlalchemy
plotly
lifelines
streamlit
dbt-postgres>=1.5.0  